@@ -5,7 +5,7 @@ namespace ETHotfix
 {
 	// 用来测试消息包含复杂类型，是否产生gc
 	[MessageHandler(AppType.Gate)]
-	public class G2C_PlayerInfoHandler : AMRpcHandler<C2G_PlayerInfo, G2C_PlayerInfo>
+	public class C2G_PlayerInfoHandler : AMRpcHandler<C2G_PlayerInfo, G2C_PlayerInfo>
 	{
 		protected override void Run(Session session, C2G_PlayerInfo message, Action<G2C_PlayerInfo> reply)
 		{
