@@ -12,6 +12,7 @@ namespace ETHotfix
         }
         
         // 每10s扫描一次过期的actorproxy进行回收,过期时间是1分钟
+        // 可能由于bug或者进程挂掉，导致ActorLocationSender发送的消息没有确认，结果无法自动删除，每一分钟清理一次这种ActorLocationSender
         public async ETVoid StartAsync(ActorLocationSenderComponent self)
         {
             List<long> timeoutActorProxyIds = new List<long>();
@@ -36,7 +37,7 @@ namespace ETHotfix
                         continue;
                     }
 
-                    if (timeNow < actorLocationMessageSender.LastSendTime + 60 * 1000)
+                    if (timeNow < actorLocationMessageSender.LastRecvTime + 60 * 1000)
                     {
                         continue;
                     }
