@@ -147,7 +147,7 @@ namespace MongoDB.Driver.Core.Operations
 
             // constructors
             /// <summary>
-            /// Initializes a new instance of the <see cref="Acknowledged" /> class.
+            /// Initializes a new instance of the class.
             /// </summary>
             /// <param name="requestCount">The request count.</param>
             /// <param name="matchedCount">The matched count.</param>
