@@ -3,112 +3,118 @@ using System.Collections.Generic;
 
 namespace ETHotfix
 {
-	public class ComponentQueue: Component
-	{
-		public string TypeName { get; }
-		
-		private readonly Queue<Component> queue = new Queue<Component>();
+    public class ComponentQueue: Component
+    {
+        public string TypeName { get; }
 
-		public ComponentQueue(string typeName)
-		{
-			this.TypeName = typeName;
-		}
+        private readonly Queue<Component> queue = new Queue<Component>();
 
-		public void Enqueue(Component component)
-		{
-			component.Parent = this;
-			this.queue.Enqueue(component);
-		}
+        public ComponentQueue(string typeName)
+        {
+            this.TypeName = typeName;
+        }
 
-		public Component Dequeue()
-		{
-			return this.queue.Dequeue();
-		}
+        public void Enqueue(Component component)
+        {
+            component.Parent = this;
+            this.queue.Enqueue(component);
+        }
 
-		public Component Peek()
-		{
-			return this.queue.Peek();
-		}
+        public Component Dequeue()
+        {
+            return this.queue.Dequeue();
+        }
 
-		public int Count
-		{
-			get
-			{
-				return this.queue.Count;
-			}
-		}
+        public Component Peek()
+        {
+            return this.queue.Peek();
+        }
 
-		public override void Dispose()
-		{
-			if (this.IsDisposed)
-			{
-				return;
-			}
-			base.Dispose();
+        public int Count
+        {
+            get
+            {
+                return this.queue.Count;
+            }
+        }
 
-			while (this.queue.Count > 0)
-			{
-				Component component = this.queue.Dequeue();
-				component.IsFromPool = false;
-				component.Dispose();
-			}
-		}
-	}
-	
+        public override void Dispose()
+        {
+            if (this.IsDisposed)
+            {
+                return;
+            }
+
+            base.Dispose();
+
+            while (this.queue.Count > 0)
+            {
+                Component component = this.queue.Dequeue();
+                component.IsFromPool = false;
+                component.Dispose();
+            }
+        }
+    }
+
     public class ObjectPool: Component
     {
-	    public string Name { get; set; }
-	    
+        public string Name { get; set; }
+
         private readonly Dictionary<Type, ComponentQueue> dictionary = new Dictionary<Type, ComponentQueue>();
 
         public Component Fetch(Type type)
         {
-	        Component obj;
+            Component obj;
             if (!this.dictionary.TryGetValue(type, out ComponentQueue queue))
             {
-	            obj = (Component)Activator.CreateInstance(type);
+                obj = (Component) Activator.CreateInstance(type);
             }
-	        else if (queue.Count == 0)
+            else if (queue.Count == 0)
             {
-	            obj = (Component)Activator.CreateInstance(type);
+                obj = (Component) Activator.CreateInstance(type);
             }
             else
             {
-	            obj = queue.Dequeue();
+                obj = queue.Dequeue();
             }
-	        
-	        obj.IsFromPool = true;
+
+            obj.IsFromPool = true;
             return obj;
         }
 
-        public T Fetch<T>() where T: Component
-		{
-            T t = (T) this.Fetch(typeof(T));
-			return t;
-		}
-        
+        public T Fetch<T>() where T : Component
+        {
+            T t = (T) this.Fetch(typeof (T));
+            return t;
+        }
+
         public void Recycle(Component obj)
         {
-	        obj.Parent = this;
+            obj.Parent = this;
             Type type = obj.GetType();
-	        ComponentQueue queue;
+            ComponentQueue queue;
             if (!this.dictionary.TryGetValue(type, out queue))
             {
                 queue = new ComponentQueue(type.Name);
-	            queue.Parent = this;
-				this.dictionary.Add(type, queue);
+                queue.Parent = this;
+#if !SERVER
+                queue.GameObject.name = $"{type.Name}s";
+#endif
+                this.dictionary.Add(type, queue);
             }
+
             queue.Enqueue(obj);
         }
 
-	    public void Clear()
-	    {
-		    foreach (var kv in this.dictionary)
-		    {
-			    kv.Value.IsFromPool = false;
-			    kv.Value.Dispose();
-		    }
-		    this.dictionary.Clear();
-	    }
+        public void Clear()
+        {
+            foreach (var kv in this.dictionary)
+            {
+                kv.Value.IsFromPool = false;
+                kv.Value.Dispose();
+            }
+
+            this.dictionary.Clear();
+        }
     }
 }
