@@ -45,14 +45,14 @@ namespace ETHotfix
 		}
 
 		public static void Close()
-		{
-			eventSystem = null;
-			
+		{	
 			scene?.Dispose();
 			scene = null;
 			
 			objectPool?.Dispose();
 			objectPool = null;
+			
+			eventSystem = null;
 		}
 	}
 }
