@@ -191,10 +191,10 @@ namespace ETModel
 			this.allComponents.Remove(instanceId);
 		}
 
-		public Component Get(long id)
+		public Component Get(long instanceId)
 		{
 			Component component = null;
-			this.allComponents.TryGetValue(id, out component);
+			this.allComponents.TryGetValue(instanceId, out component);
 			return component;
 		}
 		
