@@ -97,7 +97,7 @@ namespace ETModel
                 queue = new ComponentQueue(type.Name);
 	            queue.Parent = this;
 #if !SERVER
-	            queue.GameObject.name = type.Name;
+	            queue.GameObject.name = $"{type.Name}s";
 #endif
 				this.dictionary.Add(type, queue);
             }
