@@ -11,33 +11,24 @@ using UnityEditor;
 
 namespace ETModel
 {
+    [ObjectSystem]
+    public class ABInfoAwakeSystem: AwakeSystem<ABInfo, string, AssetBundle>
+    {
+        public override void Awake(ABInfo self, string abName, AssetBundle a)
+        {
+            self.AssetBundle = a;
+            self.Name = abName;
+            self.RefCount = 1;
+        }
+    }
+
     public class ABInfo: Component
     {
-        private int refCount;
-        public string Name { get; }
+        public string Name { get; set; }
 
-        public int RefCount
-        {
-            get
-            {
-                return this.refCount;
-            }
-            set
-            {
-                //Log.Debug($"{this.Name} refcount: {value}");
-                this.refCount = value;
-            }
-        }
+        public int RefCount { get; set; }
 
-        public AssetBundle AssetBundle { get; }
-
-        public ABInfo(string name, AssetBundle ab)
-        {
-            this.Name = name;
-            this.AssetBundle = ab;
-            this.RefCount = 1;
-            //Log.Debug($"load assetbundle: {this.Name}");
-        }
+        public AssetBundle AssetBundle { get; set; }
 
         public override void Dispose()
         {
@@ -54,6 +45,9 @@ namespace ETModel
             {
                 this.AssetBundle.Unload(true);
             }
+
+            this.RefCount = 0;
+            this.Name = "";
         }
     }
 
@@ -196,7 +190,7 @@ namespace ETModel
 
             foreach (var abInfo in this.bundles)
             {
-                abInfo.Value?.AssetBundle?.Unload(true);
+                abInfo.Value.Dispose();
             }
 
             this.bundles.Clear();
@@ -227,12 +221,13 @@ namespace ETModel
             {
                 throw new Exception($"not found bundle: {abName}");
             }
+
             return abInfo.AssetBundle;
         }
-        
+
         public void UnloadBundle(string assetBundleName)
         {
-            assetBundleName = assetBundleName.ToLower();
+            assetBundleName = assetBundleName.BundleNameToLower();
 
             string[] dependencies = AssetBundleHelper.GetSortedDependencies(assetBundleName);
 
@@ -245,7 +240,7 @@ namespace ETModel
 
         private void UnloadOneBundle(string assetBundleName)
         {
-            assetBundleName = assetBundleName.ToLower();
+            assetBundleName = assetBundleName.BundleNameToLower();
 
             ABInfo abInfo;
             if (!this.bundles.TryGetValue(assetBundleName, out abInfo))
@@ -263,6 +258,7 @@ namespace ETModel
             }
 
             this.bundles.Remove(assetBundleName);
+            this.resourceCache.Remove(assetBundleName);
             abInfo.Dispose();
             //Log.Debug($"cache count: {this.cacheDictionary.Count}");
         }
@@ -323,8 +319,7 @@ namespace ETModel
                     AddResource(assetBundleName, assetName, resource);
                 }
 
-                abInfo = new ABInfo(assetBundleName, null);
-                abInfo.Parent = this;
+                abInfo = ComponentFactory.CreateWithParent<ABInfo, string, AssetBundle>(this, assetBundleName, null);
                 this.bundles[assetBundleName] = abInfo;
                 return;
 #endif
@@ -357,8 +352,7 @@ namespace ETModel
                 }
             }
 
-            abInfo = new ABInfo(assetBundleName, assetBundle);
-            abInfo.Parent = this;
+            abInfo = ComponentFactory.CreateWithParent<ABInfo, string, AssetBundle>(this, assetBundleName, assetBundle);
             this.bundles[assetBundleName] = abInfo;
         }
 
@@ -405,10 +399,9 @@ namespace ETModel
                     AddResource(assetBundleName, assetName, resource);
                 }
 
-                abInfo = new ABInfo(assetBundleName, null);
-                abInfo.Parent = this;
+                abInfo = ComponentFactory.CreateWithParent<ABInfo, string, AssetBundle>(this, assetBundleName, null);
                 this.bundles[assetBundleName] = abInfo;
-                return;    
+                return;
 #endif
             }
 
@@ -444,8 +437,7 @@ namespace ETModel
                 }
             }
 
-            abInfo = new ABInfo(assetBundleName, assetBundle);
-            abInfo.Parent = this;
+            abInfo = ComponentFactory.CreateWithParent<ABInfo, string, AssetBundle>(this, assetBundleName, assetBundle);
             this.bundles[assetBundleName] = abInfo;
         }
 
